requirements.txt
apipkg==1.5
atomicwrites==1.4.0
attrs==20.3.0
cachetools==4.2.0
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
colorama==0.4.4
execnet==1.7.1
Flask==1.1.2
gevent==20.12.1
google-api-core==1.24.1
google-auth==1.24.0
google-cloud-core==1.5.0
google-cloud-datastore==1.15.3
google-cloud-ndb==1.7.2
googleapis-common-protos==1.52.0
greenlet==0.4.17
grpcio==1.34.0
gunicorn==20.0.4
idna==2.10
importlib-metadata==3.3.0
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
packaging==20.8
pluggy==0.13.1
protobuf==3.14.0
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pymemcache==3.4.0
pyparsing==2.4.7
pytest==6.2.1
pytest-forked==1.3.0
pytest-mock==3.4.0
pytest-xdist==2.2.0
pytz==2020.5
redis==3.5.3
requests==2.25.1
rsa==4.6
six==1.15.0
toml==0.10.2
typing-extensions==3.7.4.3
urllib3==1.26.2
Werkzeug==1.0.1
zipp==3.4.0
zope.event==4.5.0
zope.interface==5.2.0
